Move some files to more logical locations.
