+++ src/components/create_cluster/new_aws_worker.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import BootstrapModal from 'react-bootstrap/lib/Modal';
-import Button from '../button';
+import Button from '../shared/button';
 import InputField from '../shared/input_field';
 import PropTypes from 'prop-types';
