+++ src/components/home/cluster_dashboard_item.js
@@ -6,7 +6,7 @@ import ClusterIDLabel from '../shared/cluster_id_label';
 import { Link } from 'react-router-dom';
 import { relativeDate } from '../../lib/helpers.js';
 import ButtonGroup from 'react-bootstrap/lib/ButtonGroup';
-import Button from '../button/index';
+import Button from '../shared/button';
 import PropTypes from 'prop-types';
 import { push } from 'connected-react-router';
