+++ src/components/modal/release_details_modal.js
@@ -2,7 +2,7 @@
 
 import _ from 'underscore';
 import BootstrapModal from 'react-bootstrap/lib/Modal';
-import Button from '../button/index';
+import Button from '../shared/button';
 import React from 'react';
 import { relativeDate } from '../../lib/helpers.js';
 import PropTypes from 'prop-types';
